lab 3.4 Тест

+++ app/src/main/java/com/example/labs/data/lab3/FeistelNetworkKeyGenerator.kt
@@ -2,10 +2,11 @@ package com.example.labs.data.lab3
 
 import java.security.SecureRandom
 
+
 object FeistelNetworkKeyGenerator {
 
-     fun keyGenerator(): Int{
-         val secureRandom = SecureRandom()
-         return secureRandom.nextInt()
+    fun keyGenerator(): Long {
+        val secureRandom = SecureRandom()
+        return secureRandom.nextLong()
     }
 }
